Working on the Library Project
Working on the Library Project for The Odin Project.

JS is creating the cards from the Objects. Remove button is working correctly.
The Mark as Read checkbox is changing the textContent of the Card but not the actual Object value - I need to fix this.
I also need to add the New Book button and form.