Progress on Library
I have fixed the Mark as Read checkbox so that the actual array is updated. I have also started building out the form.

I need to link the Add Book button to the create book function. I also need to amend the Remove button so that it removes the object from the array.