The Odin Project - Library Project
I have completed The Odin Project's Library.

This is a page that allows the user to fill in a form to add a Book to their digital Library. At this stage the Library does not store the data, but we will be revisiting this at a later time to add this functionality.

I have added buttons allowing the user to toggle the Read status of their books. I have also added a Remove button that remove the books.

These buttons update the object data in the array and also manipulate the DOM to show the updated information based on the information in the array.